(maint) Add bind mount for Workspace and docker
When running things in k8s locally the niceties from docker desktop
aren't available, so you actually have to pass paths relative to the
docker-desktop WSL distro.

So the only way I've found to actually mount hostpath volumes with k8s
is to bind mount the workspace in the shared distro space and then use
paths relative to the docker-desktop distro.